backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.0
marshmallow-sqlalchemy==1.0.0
marshmallow==3.20.1
flask-cors==4.0.0
pytest==7.4.3